ElixirBridge v0.0.1 Mexico

Destination: Cancún, México.

Alright ! Before you catch your flight let’s talk about the client.

Mr. Sanchez is a business owner interested in providing great service at his hotels. He’s interested in expanding the experience by entering to the digital world. The plan is to run experiments to gather more data about the market. Last year, another company that worked for Mr. Sanchez, built a database in elixir that we can use !

While you’re on the air, make sure you learn about iex. It’s a tool that will let you code and read the results of your code instantly. It’s known as “Interactive Elixir” and we will be using it to access Mr. Sanchez’s database.

This guide will help you out.:w

Functions

database()